refactor: rename pipeline_mixin.py to pipeline_step_mixin.py and corresponding tests
