+++ junifer/pipeline/__init__.py
@@ -3,4 +3,4 @@
 # Authors: Synchon Mandal <s.mandal@fz-juelich.de>
 # License: AGPL
 
-from .pipeline_mixin import PipelineStepMixin
+from .pipeline_step_mixin import PipelineStepMixin
